refactor: move git specific code to a new module
